:recycle: Rename to calendar page store
