+++ lib/domain/calendar/components/pill_sheet_modified_history/pill_sheet_modified_history_card.dart
@@ -6,7 +6,7 @@ import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/components/molecules/app_card.dart';
 import 'package:pilll/components/molecules/premium_badge.dart';
-import 'package:pilll/domain/calendar/calendar_store.dart';
+import 'package:pilll/domain/calendar/calendar_page_store.dart';
 import 'package:pilll/domain/calendar/components/pill_sheet_modified_history/components/pill_sheet_modified_history_more_button.dart';
 import 'package:pilll/domain/calendar/components/pill_sheet_modified_history/pill_sheet_modified_history_list.dart';
 import 'package:pilll/domain/calendar/components/pill_sheet_modified_history/pill_sheet_modified_history_list_header.dart';
